fix: 쿨타임 float으로

+++ Assets/1. Scripts/Loader/CommonSettingsModel.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class CommonSettingsModel
 {
     [SerializeField] public List<UnitType> initalUnitlist;
-    [SerializeField] public int skillCooldown;
+    [SerializeField] public float skillCooldown;
 }
